Home

Awaits.jl

Awaits.@await

Awaits.@cancelscope

Awaits.@check

Awaits.@go

Awaits.@in

Awaits.@taskgroup

Awaits.@await — Macro.

@await body

Handle an exception returned from body. If body evaluates to an Exception, it cancels all the tasks attached to the current context.

The variable _ in code like @await f(_, x, y) is replaced by the current task context. It must be invoked inside @taskgroup or @in macros.

source

Awaits.@cancelscope — Macro.

@cancelscope(body) :: TaskContext

Create a task context with a new scope of cancellation. That is to say, error in @await inside of @cancelscope do not cancel tasks outside of this scope.

source

Awaits.@check — Macro.

@check [context]

Check context and exit the function (do return Cancelled()) if it is required.

source

Awaits.@go — Macro.

@go body

Equivalent to Threads.@spawn but it lets you pass current context (semi-)automatically.

The variable _ in code like @go f(_, x, y) is replaced by the current task context. It must be invoked inside @taskgroup or @in macros.

source

Awaits.@in — Macro.

@in context body

Enter into the task context so that other macros like @go works as expected. This should be used inside a function in which the task context is passed by a caller (for example, using @go f(_) macro).

source

Awaits.@taskgroup — Macro.

@taskgroup(body)

Create a new task context. This is an extension of the @sync block.

source